+++ src/components/BackdopApp.tsx
@@ -9,6 +9,7 @@ import {
 // ScreenGuiDom
 import React, { ReactNode, useCallback, useEffect } from "react";
 import { Engine, Scene } from "react-babylonjs";
+import { Globals } from "react-spring";
 import { toRadians } from "shutils/dist/speedAngleDistance";
 import {
   BackdopConcepFuncs,
@@ -55,6 +56,8 @@ export function makeBackdopApp<
   speechVidFiles: SpeechVidFiles
 ) {
   const { getRefs, onNextTick, setState } = concepFuncs;
+
+  Globals.assign({ frameLoop: "always", requestAnimationFrame: onNextTick });
 
   const ScreenGuiDom = makeScreenGui<
     ConcepFuncs,
@@ -121,48 +124,48 @@ export function makeBackdopApp<
         id="app"
         style={{ width: "100vw", height: "100vh", overflow: "hidden" }}
       >
+        {/*
         <div
           id="attachVidsForAutoPlay"
           style={{ display: "none", position: "absolute" }}
         ></div>
+        */}
         {/* <AllTestVideoStuff /> */}
 
-        <Engine canvasId="scene-canvas" adaptToDeviceRatio={false}>
+        <Engine
+          canvasId="scene-canvas"
+          adaptToDeviceRatio={false}
+          engineOptions={{
+            disableWebGL2Support: false,
+            powerPreference: "high-performance",
+          }}
+        >
           <Scene
             clearColor={Color4.FromColor3(Color3.FromHexString("#aca898"), 0.0)}
             // onSceneMount={(info) => (globalRefs.scenes.main = info.scene)}
             onSceneMount={(info) => {
+              info.scene.autoClear = false;
+              info.scene.autoClearDepthAndStencil = false;
+              // info.scene.blockMaterialDirtyMechanism = true;
+              // setTimeout(() => {
+              // info.scene.freezeActiveMeshes();
+              // }, 5000);
               globalRefs.scenes.main = info.scene;
+              globalRefs.scenes.backdrop = info.scene;
 
               const engine = info.scene.getEngine();
 
-              if (engine._workingCanvas) {
-                // engine._workingCanvas.width = 1280;
-                // engine._workingCanvas.height = 720;
-              }
+              // engine.setHardwareScalingLevel(8);
+              // if (engine._workingCanvas) {
+              // engine._workingCanvas.width = 1280;
+              // engine._workingCanvas.height = 720;
+              // }
 
               engine.onResizeObservable.add(() => {
                 setState({
                   global: { main: { timeScreenResized: Date.now() } },
                 });
               });
-
-              // Each frame is rendered manually inside the video looping check function atm
-              engine.stopRenderLoop();
-            }}
-          >
-            <targetCamera
-              name="camera1"
-              position={new Vector3(0, 7, -15)}
-              rotation={new Vector3(toRadians(15), toRadians(0), 0)}
-              fov={toRadians(50)}
-            />
-            <LoadingModels>{children}</LoadingModels>
-          </Scene>
-          <Scene
-            clearColor={Color4.FromColor3(Color3.FromHexString("#000000"))}
-            onSceneMount={(info) => {
-              globalRefs.scenes.backdrop = info.scene;
 
               onNextTick(() => {
                 if (globalRefs.scenes.backdrop) {
@@ -174,8 +177,14 @@ export function makeBackdopApp<
                   );
                 }
               });
+
+              // engine.setHardwareScalingLevel(1);
+              // Each frame is rendered manually inside the video looping check function atm
+              engine.stopRenderLoop();
             }}
           >
+            <LoadingModels>{children}</LoadingModels>
+
             <targetCamera
               name="camera1"
               position={new Vector3(0, 0, -2)}
@@ -183,6 +192,7 @@ export function makeBackdopApp<
               mode={Camera.ORTHOGRAPHIC_CAMERA}
               ref={scenePlaneCameraRef}
             />
+            {/*  might need to make scene plane programatically, to make sure it's not in the the scene view? */}
             <ScenePlane />
           </Scene>
         </Engine>
@@ -191,6 +201,37 @@ export function makeBackdopApp<
     );
   };
 }
+
+// <Scene
+//   clearColor={Color4.FromColor3(Color3.FromHexString("#000000"))}
+//   onSceneMount={(info) => {
+//     globalRefs.scenes.backdrop = info.scene;
+//     info.scene.autoClear = false;
+//     info.scene.autoClearDepthAndStencil = false;
+//     // info.scene.blockMaterialDirtyMechanism = true;
+//     info.scene.freezeActiveMeshes();
+//
+//     // onNextTick(() => {
+//     //   if (globalRefs.scenes.backdrop) {
+//     //     // const postProcess =
+//     //     new FxaaPostProcess(
+//     //       "fxaa",
+//     //       1.0,
+//     //       globalRefs.scenes.backdrop.activeCamera
+//     //     );
+//     //   }
+//     // });
+//   }}
+// >
+//   <targetCamera
+//     name="camera1"
+//     position={new Vector3(0, 0, -2)}
+//     rotation={new Vector3(toRadians(0), toRadians(0), 0)}
+//     mode={Camera.ORTHOGRAPHIC_CAMERA}
+//     ref={scenePlaneCameraRef}
+//   />
+//   <ScenePlane />
+// </Scene>
 
 // add this to see scene behind the scene texture rectangle
 // info.scene.autoClear = false;